chore(frontend): remove (user)/settings phase-out tree, move PasswordChangeCard to components/settings
- git mv PasswordChangeCard.tsx to frontend/components/settings/PasswordChangeCard.tsx
- Update frontend/app/(partner)/partner/settings/page.tsx import to new path
- git rm -r frontend/app/(user)/settings/ (8 files removed)
- tsc --noEmit: 0 errors; i18n check: no new missing keys
