+++ frontend/app/(partner)/partner/settings/page.tsx
@@ -22,7 +22,7 @@ import {
 import { toast } from 'sonner'
 import { useAuth } from '@/lib/auth'
 import { updateUser } from '@/lib/api/users'
-import { PasswordChangeCard } from '@/app/(user)/settings/_components/PasswordChangeCard'
+import { PasswordChangeCard } from '@/components/settings/PasswordChangeCard'
 import { getStoredToken } from '@/lib/auth'
 import { putJson } from '@/lib/api/http'
 import { WalletConnectCard } from '@/components/partner/WalletConnectCard'
