--- frontend/app/(user)/settings/_components/RiskSettingsCard.tsx
@@ -1,205 +1,0 @@
-// Copyright (c) Ultra AutoTrade. All rights reserved.
-// Unauthorized copying or distribution is strictly prohibited.
-'use client'
-
-import { useState } from 'react'
-import { Lock } from 'lucide-react'
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
-import { Input } from '@/components/ui/input'
-import { Label } from '@/components/ui/label'
-import { cn } from '@/lib/utils'
-import { toast } from 'sonner'
-import { apiPut } from '@/lib/api/client'
-
-type RiskMode = 'conservative' | 'balanced' | 'aggressive'
-
-interface RiskModeResponse {
-  mode: RiskMode
-  message: string
-}
-
-interface RiskSettingsCardProps {
-  riskMode: RiskMode
-  onRiskModeChange: (mode: RiskMode) => void
-  maxSingleTradeUsd: number
-  onMaxSingleTradeUsdChange: (value: number) => void
-  maxDailyTradeUsd: number
-  onMaxDailyTradeUsdChange: (value: number) => void
-  allowedModes?: RiskMode[]
-}
-
-const riskOptions: {
-  mode: RiskMode
-  label: string
-  emoji: string
-  description: string
-  toastDescription: string
-}[] = [
-  {
-    mode: 'conservative',
-    label: '保守的',
-    emoji: '🛡️',
-    description: '安全重視、低リスク',
-    toastDescription: 'ステーブルコインのみで安全に運用します',
-  },
-  {
-    mode: 'balanced',
-    label: 'バランス',
-    emoji: '⚖️',
-    description: '標準設定',
-    toastDescription: '安定性と収益のバランスを取って運用します',
-  },
-  {
-    mode: 'aggressive',
-    label: '積極的',
-    emoji: '🚀',
-    description: '高リターン、高リスク',
-    toastDescription: '高リターンを重視して積極的に運用します',
-  },
-]
-
-const modeActiveClass: Record<RiskMode, string> = {
-  conservative: 'border-emerald-500 bg-emerald-950/20',
-  balanced: 'border-blue-500 bg-blue-950/20',
-  aggressive: 'border-orange-500 bg-orange-950/20',
-}
-
-export function RiskSettingsCard({
-  riskMode,
-  onRiskModeChange,
-  maxSingleTradeUsd,
-  onMaxSingleTradeUsdChange,
-  maxDailyTradeUsd,
-  onMaxDailyTradeUsdChange,
-  allowedModes = ['conservative'],
-}: RiskSettingsCardProps) {
-  const [saving, setSaving] = useState(false)
-
-  const handleRiskModeChange = async (mode: RiskMode) => {
-    if (saving || mode === riskMode || !allowedModes.includes(mode)) return
-    setSaving(true)
-    try {
-      await apiPut<RiskModeResponse>('/auth/risk-mode', { mode })
-      onRiskModeChange(mode)
-      const opt = riskOptions.find((o) => o.mode === mode)
-      toast.success(`${opt?.label ?? mode}に変更しました`, {
-        description: opt?.toastDescription,
-      })
-    } catch {
-      toast.error('リスクモードの保存に失敗しました')
-    } finally {
-      setSaving(false)
-    }
-  }
-
-  return (
-    <Card className="bg-zinc-900 border-zinc-800">
-      <CardHeader className="pb-3">
-        <CardTitle className="text-base text-zinc-100">リスク設定</CardTitle>
-      </CardHeader>
-      <CardContent className="space-y-5">
-        {/* リスクモード選択 */}
-        <div className="space-y-2">
-          <Label className="text-zinc-300 text-sm">リスクモード</Label>
-          <div className="grid grid-cols-1 gap-2 sm:grid-cols-3">
-            {riskOptions.map((opt) => {
-              const isActive = riskMode === opt.mode
-              const isAllowed = allowedModes.includes(opt.mode)
-              const isDisabled = saving || !isAllowed
-
-              return (
-                <button
-                  key={opt.mode}
-                  onClick={() => { void handleRiskModeChange(opt.mode) }}
-                  disabled={isDisabled}
-                  className={cn(
-                    'text-left rounded-lg border-2 p-3 transition-all',
-                    isAllowed
-                      ? 'disabled:opacity-60'
-                      : 'cursor-not-allowed opacity-50',
-                    isActive && isAllowed
-                      ? modeActiveClass[opt.mode]
-                      : isAllowed
-                        ? 'border-zinc-700 bg-zinc-800/40 hover:border-zinc-600'
-                        : 'border-zinc-800 bg-zinc-900/40',
-                  )}
-                >
-                  <div className="flex items-center gap-1.5 mb-1">
-                    {isAllowed ? (
-                      <span className="text-base">{opt.emoji}</span>
-                    ) : (
-                      <Lock className="w-4 h-4 text-zinc-600" />
-                    )}
-                    <span
-                      className={cn(
-                        'text-sm font-semibold',
-                        isAllowed ? 'text-zinc-100' : 'text-zinc-600',
-                      )}
-                    >
-                      {opt.label}
-                    </span>
-                    {isActive && isAllowed && (
-                      <span className="ml-auto text-xs px-1.5 py-0.5 rounded bg-zinc-700 text-zinc-300">
-                        現在
-                      </span>
-                    )}
-                    {!isAllowed && (
-                      <span className="ml-auto text-xs px-1.5 py-0.5 rounded bg-zinc-800 text-zinc-600">
-                        Phase 2
-                      </span>
-                    )}
-                  </div>
-                  <p className={cn('text-xs', isAllowed ? 'text-zinc-400' : 'text-zinc-700')}>
-                    {isAllowed ? opt.description : 'Phase 2 で解禁予定'}
-                  </p>
-                </button>
-              )
-            })}
-          </div>
-        </div>
-
-        {/* 最大単回取引額 */}
-        <div className="space-y-1.5">
-          <Label htmlFor="max-single-trade" className="text-zinc-300 text-sm">
-            最大単回取引額 (USD)
-          </Label>
-          <Input
-            id="max-single-trade"
-            type="number"
-            min={1}
-            step={10}
-            value={maxSingleTradeUsd}
-            onChange={(e) => {
-              onMaxSingleTradeUsdChange(Number(e.target.value))
-              // TODO: Save to backend PUT /api/user/settings
-            }}
-            placeholder="500"
-            className="bg-zinc-800 border-zinc-700 text-zinc-100 placeholder:text-zinc-500"
-          />
-        </div>
-
-        {/* 日次取引上限 */}
-        <div className="space-y-1.5">
-          <Label htmlFor="max-daily-trade" className="text-zinc-300 text-sm">
-            日次取引上限 (USD)
-          </Label>
-          <Input
-            id="max-daily-trade"
-            type="number"
-            min={1}
-            step={100}
-            value={maxDailyTradeUsd}
-            onChange={(e) => {
-              onMaxDailyTradeUsdChange(Number(e.target.value))
-              // TODO: Save to backend PUT /api/user/settings
-            }}
-            placeholder="2000"
-            className="bg-zinc-800 border-zinc-700 text-zinc-100 placeholder:text-zinc-500"
-          />
-        </div>
-
-        <p className="text-xs text-zinc-500">* 管理者設定の範囲内で調整できます</p>
-      </CardContent>
-    </Card>
-  )
-}
