--- frontend/app/(user)/settings/_components/index.ts
@@ -1,9 +1,0 @@
-// Copyright (c) Ultra AutoTrade. All rights reserved.
-// Unauthorized copying or distribution is strictly prohibited.
-export { OperationModeCard } from './OperationModeCard'
-export { RiskSettingsCard } from './RiskSettingsCard'
-export { NotificationCard } from './NotificationCard'
-export { LanguageCard } from './LanguageCard'
-export { WalletInfoCard } from './WalletInfoCard'
-export { AppInfoCard } from './AppInfoCard'
-export { PasswordChangeCard } from './PasswordChangeCard'
